requirements.txt
click==8.0.3
fastapi==0.72.0
protobuf==3.19.3
requests==2.25.0
python-snappy==0.6.0
uvicorn==0.17.0
xmltodict==0.12.0
boto3==1.20.40
schedule==1.1.0
paho-mqtt==1.6.1